Fix and put into folders
