+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require('./getAll')
+const getContact = require('./getContactById')
+const addContact = require('./addContact')
+const deleteContact = require('./deleteContact')
+const updateContact = require('./updateContact')
+
+module.exports = {
+  getAll,
+  getContact,
+  addContact,
+  deleteContact,
+  updateContact,
+}
