+++ controllers/index.js
@@ -1,0 +1,5 @@
+const contactsController = require('./contacts')
+
+module.exports = {
+  contactsController,
+}
