+++ db/contacts.json
@@ -1,0 +1,1 @@
+[{"id":1,"name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":2,"name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":3,"name":"Kennedy Lane","email":"mattis.Cras@nonenimMauris.net","phone":"(542) 451-7038"},{"id":4,"name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":5,"name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":6,"name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":7,"name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":8,"name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":9,"name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":10,"name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"name":"Natasha Totska","email":"natasha.totska@yahoo.com","phone":"(097) 111-7190","id":"1ccdb9e2-5e1f-4621-bee0-2fc8fed51ece"},{"name":"Yuriy Totskyy","email":"yuriy@yahoo.com","phone":"(097) 111-7140","id":"c81400e9-1e88-40c1-80c7-bd697273747f"}]
