+++ model/contacts/index.js
@@ -1,0 +1,13 @@
+const listContacts = require("./listContacts");
+const getContactById = require("./getContactById");
+const updateContact = require("./updateContact");
+const removeContact = require("./removeContact");
+const addContact = require("./addContact");
+
+module.exports = {
+  listContacts,
+  getContactById,
+  updateContact,
+  removeContact,
+  addContact,
+};
