+++ routes/api/contacts.js
@@ -1,110 +1,16 @@
 const express = require("express");
-const contactsOperations = require("../../model/index");
 const router = express.Router();
-const { contactSchema } = require("../../schemas");
 
-router.get("/", async (req, res, next) => {
-  try {
-    const contacts = await contactsOperations.listContacts();
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        result:contacts,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+const { contactsController } = require("../../controllers");
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsOperations.getContactById(contactId);
-    if (!result) {
-      const error = new Error(`Contact with id=${contactId} not found`);
-      error.status = 404;
-      throw error;
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        result,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", contactsController.getAll);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactSchema.validate(req.body);
-    if (error) {
-      const err = new Error(error.message);
-      err.status = 400;
-      throw err;
-    }
-    const result = await contactsOperations.addContact(req.body);
-    res.status(201).json({
-      status: "success",
-      code: "201",
-      data: {
-        result
-      }
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", contactsController.getContact);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsOperations.removeContact(contactId);
-    if (!result) {
-      const error = new Error("Not found");
-      error.status = 404;
-      throw error;
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      message: "contact deleted"
-    })
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", contactsController.addContact);
 
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = contactSchema.validate(req.body);
-    if (error) {
-      const err = new Error("Missing fields");
-      err.status = 400;
-      throw err;
-    }
-    const { contactId } = req.params;
-    const result = await contactsOperations.updateContact(contactId, req.body);
-    if (!result) {
-      const error = new Error("Not found");
-      error.status = 404;
-      throw error;
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        result,
-      },
-    });
+router.delete("/:contactId", contactsController.deleteContact);
 
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", contactsController.updateContact);
 
 module.exports = router;
